Resolve int to float bit conversion and retrieval, to render in-memory programs in color.

+++ src/edu/smu/cse7343/bakos/os/Memory.java
@@ -9,17 +9,17 @@ package edu.smu.cse7343.bakos.os;
 
 public class Memory {
 
-    int[] registers;
+    float[] registers;
 
     public Memory(int size) {
-        registers = new int[size];
+        registers = new float[size];
     }
 
     public int totalSize() {
         return registers.length;
     }
 
-    public void write(int address, int value) {
+    public void write(int address, float value) {
         if (address >= registers.length) {
             System.out.println("TRAP: Address out of bounds."); // TODO
         } else {
